Refactored vispy.util.testing
Basically, the testing module is not imported by default. Stuff that
we need is moved out of it. Much test-related stuff is moved into it.

* app_opengl_context -> moved to vispy.app._util
* _TempDir -> moved to util._config
* requires_* -> moved into util.testing
* Fixed sys_info() function (we had two versions omg)
